Add OpenStack policy integration bits
This patch adds the openstack policy integration bits which we can use to
enforce which congress api calls can be made by a normal tenant
vs admin tenant. Today the default policy.json defaults to requiring an
admin to make an API call. Though, for example, one could add the following
to policy.json to allow a non-admin tenant to create a data-source:

	"create_data_sources": "rule:regular_user"

This patch also renames Context to RequestContext and drops some unneeded code.

Related to blueprint keystone-integration

Change-Id: I7fe17938ba6814df3679d9d9fd0ed1cd4fc3771f